Added media commands and more
Media commands when run by a client are sent to the server so server-sided commands can be run by clients.
The GFX directory has appeared, and with it are two fancy outline effects. The first, outline.lua is a simple thick outline. Expect changes to how it renders entities soon. The second, outline_blip.lua is a pulsing outline. This effect should be used sparingly, as it is almost as expensive as rendering halos.
Localizations and several related functions have been added.
The slay and bakse command have been added, both of which are media commands.
A player discovery library has been added.
The help command has been improved to accomadate for media commands.
The COMMAND structure now automatically adds a Command key that contains the key of the structure in the global PYRITION.Commands table.
And more, less significant things.